% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupMeanSummarization.R
\name{Group_mean_summarization}
\alias{Group_mean_summarization}
\title{: Group Mean Summarization}
\usage{
Group_mean_summarization(dataFrame, group_by)
}
\arguments{
\item{dataFrame}{dataframe_like_object accepts either dataframe or tibbular data & hence the chosen name to indicate the inclusion of both}

\item{group_by}{vector should be non-empty and can either be numerical or string thus the chosen name is kept generic}
}
\value{
tibble for the case where legal arguments are passed otherwise null along with an error message
}
\description{
The function automates the calculation of arithmetic mean - operating and summarizing on the grouping chosen by the user
}
\details{
This is a generic function that can group by one or multiple levels passed by the user.
User can pass provide argument for grouping either by numerical index or column name because both ways are used in practice.
Rows with N/A values are kept as it can still be valid grouping combinations. Thus N/A values are treated via omission in the calculation of mean.
}
\examples{
sample_data <- data.frame(Name = c("Vir", "Shreya", "Kabir", "Vir", "Kabir"), TransactionCity = c("Vancouver", "Vancouver", "California", "Seattle", "California"), TransactionCost = c(100, 50, 80, 100, 40), ItemsPurchased = c(4,1,4,6,2)) # Generating sample data
Group_mean_summarization(sample_data, "Name") #Providing arithmetic mean summary per person using column name
Group_mean_summarization(sample_data, c(2))   #Providing arithmetic mean summary per city using column index
Group_mean_summarization(sample_data, c(2,1)) #Providing arithmetic mean summary per city and name using column index

}
